backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
mysql-connector-python==8.2.0
bcrypt==4.1.1
python-dotenv==1.0.0
geopy==2.4.1

